requirements.txt
paramiko
requests
argparse
colorama
PyYAML